Social comp renamed to ExternalLink + refactored
